refactor: split module_resolver.rs validation into submodule
Extract the validate/require expression evaluator (~425 lines) from
module_resolver.rs into module_resolver/validation.rs.

This separates the mini-language interpreter (validate_expr,
require_expr evaluation) from the core module loading/expansion logic.

No logic changes - pure structural refactoring. All 35 module_resolver
tests pass unchanged.

Refs #1850
